// src/global/types.ts
export type GlobalCategoryId = 'fashion' | 'electronics' | 'beauty' | 'home' | 'kids' | 'sports' | 'automotive' | 'books' | 'pets' | 'accessories';
// ...
export type ImporterModel='CUSTOMER_IMPORTER'|'KAREEBU_IMPORTER'|'THIRD_PARTY_IMPORTER'|'CONSOLIDATOR_MODEL';